fix double import, rename files
